Speed up checkdegree by handling natural An/Sn
Testing whether a permutation group is a natural An or Sn is very fast.
Testing one of these group for primitivity can be extremely slow in contrast
(depending on the degree and the random seed, we are talking about several
orders of magnitude),

By adding dedicated checks for these cases, `sanity4095.tst` now finishes
in around two minutes on my laptop. As such I promoted it to teststandard,
and removed sanity999.tst entirely.